thesis/requirements.txt
# System Requirements
# Install with: pip install -r requirements.txt

# Core Dependencies
requests>=2.31.0
python-docx>=0.8.11

# Data Analysis (for Chapter 4)
pandas>=2.0.0
openpyxl>=3.1.0

# Optional but recommended
matplotlib>=3.7.0
seaborn>=0.12.0
